apps/api/pyproject.toml
[project]
name = "quant-cover-api"
version = "0.1.0"
description = "Quant Cover API"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.2,<2.0.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "psycopg[binary]>=3.2.1,<4.0.0",
    "sqlalchemy>=2.0.35,<3.0.0",
    "uvicorn>=0.30.0,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.2,<1.0.0",
    "pre-commit>=4.2.0,<5.0.0",
    "pytest>=8.3.5,<9.0.0",
    "ruff>=0.11.5,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quant_cover_api"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 130

[tool.ruff.lint]
select = ["E", "F", "I"]
